Rename map_ok to map
This is based on the premise that the ok-value of a Try may appear to
mutate, but the error-value of a Try may not. Thus, much like a Maybe
type, we can only map over the value of one element of the sum, so we
can shorten the name.

Additionally, mapping the ok-value of a Try is the "common" use case,
and if in the future we discover a need to map the error-value, we can
burden it with a longer name like `map_error` without much harm.