Rework into functor using of, map, and reduce only.
This is to address the concern that it makes no sense to "return" a Try
across fucntion boundaries, since tags are very context-sensitive, by
design. Try is a computational counterpart to what other languages call
a Result, and what Elixir does not have a dedicated type for. So authors
would use Result as a value to pass between functions, and Try as a
computation and composition tool.